+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [1.0.8]
 - Add disable state style to workspaces ui button
+- Accept ::f.portal/root-node-props on Fulcro portal
 
 ## [1.0.7]
 - Now when you send initial state and the Fulcro component doens't implemenet it, params are used as is instead of crashing